backend/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-cors
PyJWT
requests
python-dotenv